backend/users/api/views.py
from rest_framework import generics
from users.models import Profile
from .serializers import ProfileSerializer


class Profiles(generics.ListAPIView):
    queryset = Profile.objects.all()
    serializer_class = ProfileSerializer


class UserProfile(generics.RetrieveAPIView):
    queryset = Profile.objects.all()
    serializer_class = ProfileSerializer
