backend/requirements/base.txt
appdirs==1.4.4
arrow==1.2.1
asgiref==3.4.1
astroid==2.5.2
autopep8==1.5.6
bcrypt==3.2.0
binaryornot==0.4.4
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.1
colorama==0.4.4
cookiecutter==1.7.3
cryptography==36.0.0
defusedxml==0.7.1
distlib==0.3.1
dj-rest-auth==2.1.11
Django==3.2.9
django-allauth==0.46.0
django-cors-headers==3.10.0
django-dotenv==1.4.2
django-environ==0.8.1
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
dnspython==2.1.0
email-validator==1.1.2
filelock==3.0.12
Flask==2.0.1
flask-bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
greenlet==1.1.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==3.10.0
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
jinja2-time==0.2.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
Pillow==8.2.0
poyo==0.5.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.7.4
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.12.0
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.3
redshift-sqlalchemy==0.2
requests==2.26.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.8
sqlparse==0.4.2
stripe==2.64.0
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==4.0.0
urllib3==1.26.7
virtualenv==20.4.3
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.1
